quickjs_runtime::quickjs_utils

Module compile

Utils to compile script to bytecode and run script from bytecode

Functions

compile⚠

compile a script, will result in a JSValueRef with tag JS_TAG_FUNCTION_BYTECODE or JS_TAG_MODULE. It can be executed with run_compiled_function().

from_bytecode⚠

read a compiled function from bytecode, see to_bytecode for an example

run_compiled_function⚠

run a compiled function, see compile for an example

to_bytecode⚠

write a function to bytecode