Module quickjs_runtime::quickjs_utils::functions

utils to create and invoke functions

Functions

call_constructor⚠
call a constructor (instantiate an Object)

call_constructor_q
call a constructor (instantiate an Object)

call_function⚠
call a function

call_function_q
call a function

call_to_string⚠
call an objects to_String method or convert a value to string

call_to_string_q
call an objects to_String method or convert a value to string

invoke_member_function⚠
Safety

invoke_member_function_q

is_constructor⚠
see if an Object is an instance of Function and is a constructor (can be instantiated with new keyword)

is_constructor_q
see if an Object is an instance of Function and is a constructor (can be instantiated with new keyword)

is_function⚠
see if an Object is an instance of Function

is_function_q
see if an Object is an instance of Function

new_function⚠
create a new Function which is backed by a closure

new_function_q
create a new Function which is backed by a closure

new_native_function⚠
create a new Function object which calls a native method

new_native_function_data⚠
create a new Function object which calls a native method (with data)

new_native_function_data_q
create a new Function object which calls a native method (with data)

new_native_function_q
create a new Function object which calls a native method

parse_function⚠
parse a function body and its arg_names into a JSValueRef which is a Function