quickjs_runtime::quickjs_utils::functions

Function new_native_function_data

pub unsafe fn new_native_function_data(
    context: *mut JSContext,
    func: JSCFunctionData,
    name: &str,
    arg_count: i32,
    data: QuickJsValueAdapter,
) -> Result<QuickJsValueAdapter, JsError>

create a new Function object which calls a native method (with data)

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid