Module quickjs_runtime::quickjs_utils::maps

Map utils, these methods can be used to manage Map objects from rust see MDN for more on Maps

Functions

clear⚠
remove all entries from a map

clear_q
remove all entries from a map

delete⚠
delete a value from a map by key

delete_q
delete a value from a map by key

entries⚠
iterate over all entries of a map

entries_q
iterate over all entries of a map

get⚠
get a value from a map by key

get_q
get a value from a map by key

has⚠
check whether a Map has a value for a key

has_q
check whether a Map has a value for a key

is_map⚠
see if a JSValueRef is an instance of Map

is_map_q
see if a JSValueRef is an instance of Map

keys⚠
iterate over all keys of a map

keys_q
iterate over all keys of a map

new_map⚠
create new instance of Map

new_map_q
create new instance of Map

set⚠
set a key/value pair in a Map

set_q
set a key/value pair in a Map

size⚠
get the number of entries in a map

size_q
get the number of entries in a map

values⚠
iterate over all values of a map

values_q
iterate over all values of a map