quickjs_runtime::quickjs_utils
Module typedarrays

this module is a work in progress and is currently used by me to pass Vec’s from rust to js and back again

Constants

BUFFERS

Functions

detach_array_buffer_buffer⚠

detach the array buffer and return it, after this the TypedArray is no longer usable in JS (or at least all items will return undefined)

detach_array_buffer_buffer_q

detach the array buffer and return it, after this the TypedArray is no longer usable in JS (or at least all items will return undefined)

get_array_buffer⚠

get the underlying ArrayBuffer of a TypedArray

get_array_buffer_buffer_copy⚠

Get a copy of the underlying array buffer and return it unlike when using detach_array_buffer_buffer_q the TypedArray is still intact after using this the operation is just more expensive because the Vec is cloned

get_array_buffer_buffer_copy_q

Get a copy of the underlying array buffer and return it unlike when using detach_array_buffer_buffer_q the TypedArray is still intact after using this the operation is just more expensive because the Vec is cloned

get_array_buffer_q

get the underlying ArrayBuffer of a TypedArray

is_array_buffer⚠

check if a ref is an ArrayBuffer

is_array_buffer_q

is_typed_array⚠

check if a ref is a TypedArray

is_typed_array_q

new_array_buffer⚠

this method creates a new ArrayBuffer which is used as a basis for all typed arrays the buffer vec is stored and used in js, when it is no longer needed it is dropped

new_array_buffer_copy⚠

create an array buffer with a copy of the data in a Vec

new_array_buffer_copy_q

create an array buffer with a copy of the data in a Vec

new_array_buffer_q

this method creates a new ArrayBuffer which is used as a basis for all typed arrays the buffer vec is stored and used in js, when it is no longer needed it is dropped

new_uint8_array⚠

create a new TypedArray with a buffer, the buffer is consumed and can be reclaimed later by calling detach_array_buffer_buffer_q

new_uint8_array_copy⚠

create a new TypedArray with a buffer, the buffer is copied and that copy can be reclaimed later by calling detach_array_buffer_buffer_q

new_uint8_array_copy_q

create a new TypedArray with a buffer, the buffer is copied and that copy can be reclaimed later by calling detach_array_buffer_buffer_q

new_uint8_array_q

create a new TypedArray with a buffer, the buffer is consumed and can be reclaimed later by calling detach_array_buffer_buffer_q