Struct quickjs_runtime::quickjsruntime::QuickJsRuntime

pub struct QuickJsRuntime { /* fields omitted */ }

Implementations

impl QuickJsRuntime

pub fn add_context_init_hook<H>(&self, hook: H) -> Result<(), JsError> where
H: Fn(&QuickJsRuntime, &QuickJsContext) -> Result<(), JsError> + 'static,

pub fn create_context(id: &str) -> Result<(), JsError>

pub fn remove_context(id: &str)

pub fn get_context(&self, id: &str) -> &QuickJsContext

pub fn opt_context(&self, id: &str) -> Option<&QuickJsContext>

pub fn has_context(&self, id: &str) -> bool

pub unsafe fn get_quickjs_context(
&self,
context: *mut JSContext
) -> &QuickJsContext

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid

pub fn get_rt_ref(&self) -> Option<Arc<EsRuntime>>

pub fn set_interrupt_handler<I: Fn(&QuickJsRuntime) -> bool + 'static>(
&mut self,
interrupt_handler: I
) -> &mut Self

pub fn add_script_module_loader(&mut self, sml: ScriptModuleLoaderAdapter)

pub fn add_native_module_loader(&mut self, nml: NativeModuleLoaderAdapter)

pub fn get_main_context(&self) -> &QuickJsContext

pub fn with_all_module_loaders<C, R>(&self, consumer: C) -> Option<R> where
C: Fn(&dyn ModuleLoader) -> Option<R>,

pub fn gc(&self)

run the garbage collector

pub fn do_with<C, R>(task: C) -> R where
C: FnOnce(&QuickJsRuntime) -> R,

pub fn do_with_mut<C, R>(task: C) -> R where
C: FnOnce(&mut QuickJsRuntime) -> R,

pub fn run_pending_jobs_if_any(&self)

run pending jobs if avail

todo

move this to a quickjs_utils::pending_jobs so it can be used without doing QuickjsRuntime.do_with()

pub fn has_pending_jobs(&self) -> bool

pub fn run_pending_job(&self) -> Result<(), JsError>

pub fn get_id(&self) -> &str

pub fn load_module_script_opt(
&self,
ref_path: &str,
path: &str
) -> Option<Script>

this method tries to load a module script using the runtimes script_module loaders