quickjs_runtime::reflection::eventtarget

Function dispatch_event

pub fn dispatch_event(
    q_ctx: &QuickJsRealmAdapter,
    proxy: &Proxy,
    instance_id: usize,
    event_id: &str,
    event: QuickJsValueAdapter,
) -> Result<bool, JsError>

dispatch an Event on an instance of a Proxy class the return value is false if event is cancelable and at least one of the event listeners which received event called Event.preventDefault. Otherwise it returns true