Struct quickjs_runtime::valueref::JSValueRef

pub struct JSValueRef { /* fields omitted */ }

Implementations

impl JSValueRef

pub fn get_tag(&self) -> i64

pub fn new_no_context(value: JSValue, label: &str) -> Self

pub fn new(
context: *mut JSContext,
value: JSValue,
ref_ct_incr: bool,
ref_ct_decr_on_drop: bool,
label: &str
) -> Self

pub fn get_ref_count(&self) -> i32

pub fn clone_value_incr_rc(&self) -> JSValue

borrow the value but first increment the refcount, this is useful for when the value is returned or passed to functions

pub fn borrow_value(&self) -> &JSValue

pub fn borrow_value_mut(&mut self) -> &mut JSValue

pub fn is_null_or_undefined(&self) -> bool

pub fn is_undefined(&self) -> bool

return true if the wrapped value represents a JS null value

pub fn is_null(&self) -> bool

return true if the wrapped value represents a JS null value

pub fn is_bool(&self) -> bool

return true if the wrapped value represents a JS boolean value

pub fn is_i32(&self) -> bool

return true if the wrapped value represents a JS INT value

pub fn is_module(&self) -> bool

return true if the wrapped value represents a Module

pub fn is_compiled_function(&self) -> bool

return true if the wrapped value represents a compiled function

pub fn is_f64(&self) -> bool

return true if the wrapped value represents a JS F64 value

pub fn is_big_int(&self) -> bool

pub fn is_exception(&self) -> bool

return true if the wrapped value represents a JS Exception value

pub fn is_object(&self) -> bool

return true if the wrapped value represents a JS Object value

pub fn is_string(&self) -> bool

return true if the wrapped value represents a JS String value