quickjs_runtime::values

Enum JsValueFacade

pub enum JsValueFacade {

    I32 {
        val: i32,
    },
    F64 {
        val: f64,
    },
    String {
        val: DefaultAtom,
    },
    Boolean {
        val: bool,
    },
    JsObject {
        cached_object: CachedJsObjectRef,
    },
    JsPromise {
        cached_promise: CachedJsPromiseRef,
    },
    JsArray {
        cached_array: CachedJsArrayRef,
    },
    JsFunction {
        cached_function: CachedJsFunctionRef,
    },
    Object {
        val: HashMap<String, JsValueFacade>,
    },
    Array {
        val: Vec<JsValueFacade>,
    },
    Promise {
        producer: DebugMutex<Option<Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>> + Send + 'static>>>>,
    },
    Function {
        name: String,
        arg_count: u32,
        func: Arc<Box<dyn Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync>>,
    },
    JsError {
        val: JsError,
    },
    ProxyInstance {
        namespace: &'static [&'static str],
        class_name: &'static str,
        instance_id: JsProxyInstanceId,
    },
    TypedArray {
        buffer: Vec<u8>,
        array_type: TypedArrayType,
    },
    JsonStr {
        json: String,
    },
    SerdeValue {
        value: Value,
    },
    Null,
    Undefined,

}

The JsValueFacade is a Send-able representation of a value in the Script engine

Variants

I32

Fields

val: i32

F64

Fields

val: f64

String

Fields

val: DefaultAtom

Boolean

Fields

val: bool

JsObject

Fields

cached_object: CachedJsObjectRef

JsPromise

Fields

cached_promise: CachedJsPromiseRef

JsArray

Fields

cached_array: CachedJsArrayRef

JsFunction

Fields

cached_function: CachedJsFunctionRef

Object

Fields

val: HashMap<String, JsValueFacade>

Array

Fields

val: Vec<JsValueFacade>

Promise

Fields

producer: DebugMutex<Option<Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>> + Send + 'static>>>>

Function

Fields

name: String

arg_count: u32

func: Arc<Box<dyn Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync>>

JsError

Fields

val: JsError

ProxyInstance

Fields

namespace: &'static [&'static str]

class_name: &'static str

instance_id: JsProxyInstanceId

TypedArray

Fields

buffer: Vec<u8>

array_type: TypedArrayType

JsonStr

Fields

json: String

SerdeValue

Fields

value: Value

Null

Undefined

Implementations

impl JsValueFacade

pub fn from_serializable<T: Serialize>(obj: &T) -> Result<Self, Box<dyn Error>>

pub fn new_i32(val: i32) -> Self

pub fn new_f64(val: f64) -> Self

pub fn new_bool(val: bool) -> Self

pub fn new_str_atom(val: DefaultAtom) -> Self

pub fn new_str(val: &str) -> Self

pub fn new_string(val: String) -> Self

pub fn new_callback<F: Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync + 'static>( callback: F, ) -> Self

pub fn new_function<F: Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync + 'static>( name: &str, function: F, arg_count: u32, ) -> Self

pub fn new_promise<R, P, M>(producer: P) -> Self
where P: Future<Output = Result<JsValueFacade, JsError>> + Send + 'static,

create a new promise with a producer which will run async in a threadpool

pub fn is_i32(&self) -> bool

pub fn is_f64(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_string(&self) -> bool

pub fn is_js_promise(&self) -> bool

pub fn is_js_object(&self) -> bool

pub fn is_js_array(&self) -> bool

pub fn get_i32(&self) -> i32

pub fn get_f64(&self) -> f64

pub fn get_bool(&self) -> bool

pub fn get_str(&self) -> &str

pub fn get_str_atom(&self) -> &DefaultAtom

pub fn is_null_or_undefined(&self) -> bool

pub fn get_value_type(&self) -> JsValueType

pub fn stringify(&self) -> String

pub async fn to_serde_value(&self) -> Result<Value, JsError>

pub async fn to_json_string(&self) -> Result<String, JsError>