quickjs_runtime

Module jsutils

This contains abstract traits and structs for use with different javascript runtimes the Adapter traits are use in the worker thread (EventLoop) of the Runtime and thus are not Send, they should never leave the thread The facade classes are for use outside the worker thread, they are Send

Modules

helper_tasks

jsproxies

modules

promises

Structs

JsError

Script

Enums

JsValueType

the JsValueType represents the type of value for a JSValue

Traits

ScriptPreProcessor