quickjs_runtime

Module quickjs_utils

low level contains utils for calling the quickjs api

Modules

arrays

atoms

bigints

class_ids

compile

Utils to compile script to bytecode and run script from bytecode

dates

Utils for working with Date objects

errors

utils for getting and reporting exceptions

functions

utils to create and invoke functions

interrupthandler

iterators

utils for the iterator protocol

json

serialize and stringify JavaScript objects

maps

Map utils, these methods can be used to manage Map objects from rust see MDN for more on Maps

modules

utils for working with ES6 Modules

objects

Utils for working with objects

primitives

promises

properties

runtime

sets

Set utils, these methods can be used to manage Set objects from rust see MDN for more on Sets

typedarrays

this module is a work in progress and is currently used by me to pass Vec’s from rust to js and back again

Functions

gc

get_constructor⚠

Safety

get_global⚠

Safety

get_global_q

get_memory_usage⚠

Calculate a runtimes memory usage

get_script_or_module_name⚠

get the current filename

get_script_or_module_name_q

get the current filename

new_null

new_null_ref

new_undefined

new_undefined_ref

parse_args⚠

Safety