quickjs_runtime::quickjs_utils

Module objects

Utils for working with objects

Functions

construct_object⚠

construct a new instance of a constructor

create_object⚠

create a new simple object, e.g. let obj = {};

create_object_q

create a new simple object, e.g. let obj = {};

define_getter_setter⚠

define a getter/setter property

define_getter_setter_q

define a getter/setter property

get_namespace⚠

Safety

get_namespace_q

get a namespace object this is used to get nested object properties which are used as namespaces

get_own_property_names⚠

get the property names of an object

get_own_property_names_q

get the property names of an object

get_property⚠

get a property from an object by name

get_property_names⚠

get the names of all properties of an object

get_property_names_q

get the names of all properties of an object

get_property_q

get a property from an object by name

get_prototype_of⚠

Object.prototypeOf

get_prototype_of_q

is_instance_of⚠

Safety

is_instance_of_by_name⚠

Safety

is_instance_of_by_name_q

is_instance_of_q

set_property⚠

set a property in an object, like obj[propName] = val;

set_property2⚠

set a property with specific flags

set_property2_q

set a property with specific flags set_property applies the default flag JS_PROP_C_W_E (Configurable, Writable, Enumerable) flags you can use here are

set_property_q

set a property in an object, like obj[propName] = val;

traverse_properties⚠

Safety

traverse_properties_mut⚠

Safety

traverse_properties_q

traverse_properties_q_mut