Struct quickjs_runtime::quickjsrealmadapter::QuickJsRealmAdapter

pub struct QuickJsRealmAdapter {
    pub id: String,
    pub context: *mut JSContext,
    // some fields omitted
}

Fields
id: Stringcontext: *mut JSContext
Implementations

impl QuickJsRealmAdapter

pub unsafe fn get_id(context: *mut JSContext) -> &'static str

get the id of a QuickJsContext from a JSContext

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn call_function(
&self,
namespace: Vec<&str>,
func_name: &str,
arguments: Vec<JSValueRef>
) -> Result<JSValueRef, JsError>

call a function by namespace and name

pub fn eval(&self, script: Script) -> Result<JSValueRef, JsError>

evaluate a script

pub unsafe fn eval_ctx(
context: *mut JSContext,
script: Script
) -> Result<JSValueRef, JsError>

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn eval_module(&self, script: Script) -> Result<JSValueRef, JsError>

evaluate a Module

pub unsafe fn eval_module_ctx(
context: *mut JSContext,
script: Script
) -> Result<JSValueRef, JsError>

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn report_ex(&self, err: &str) -> JSValue

throw an internal error to quickjs and create a new ex obj

pub unsafe fn report_ex_ctx(context: *mut JSContext, err: &str) -> JSValue

throw an Error in the runtime and init an Exception JSValue to return

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn get_exception_ctx(&self) -> Option<JsError>

Get the last exception from the runtime, and if present, convert it to a JsError.

pub unsafe fn get_exception(context: *mut JSContext) -> Option<JsError>

Get the last exception from the runtime, and if present, convert it to a JsError.

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid

pub fn cache_object(&self, obj: JSValueRef) -> i32

pub fn remove_cached_obj_if_present(&self, id: i32)

pub fn consume_cached_obj(&self, id: i32) -> JSValueRef

pub fn with_cached_obj<C, R>(&self, id: i32, consumer: C) -> R where
C: FnOnce(JSValueRef) -> R,

pub unsafe fn with_context<C, R>(context: *mut JSContext, consumer: C) -> R where
C: FnOnce(&QuickJsRealmAdapter) -> R,

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid