quickjs_runtime::quickjsrealmadapter

Struct QuickJsRealmAdapter

pub struct QuickJsRealmAdapter {
    pub id: String,
    pub context: *mut JSContext,
    /* private fields */
}

Fields
id: Stringcontext: *mut JSContext
Implementations

impl QuickJsRealmAdapter

pub fn print_stats(&self)

pub unsafe fn get_id(context: *mut JSContext) -> &'static str

get the id of a QuickJsContext from a JSContext

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn invoke_function_by_name( &self, namespace: &[&str], func_name: &str, arguments: &[QuickJsValueAdapter], ) -> Result<QuickJsValueAdapter, JsError>

invoke a function by namespace and name

pub fn eval(&self, script: Script) -> Result<QuickJsValueAdapter, JsError>

evaluate a script

pub fn eval_this( &self, script: Script, this: QuickJsValueAdapter, ) -> Result<QuickJsValueAdapter, JsError>

pub unsafe fn eval_ctx( context: *mut JSContext, script: Script, this_opt: Option<QuickJsValueAdapter>, ) -> Result<QuickJsValueAdapter, JsError>

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn eval_module( &self, script: Script, ) -> Result<QuickJsValueAdapter, JsError>

evaluate a Module

pub unsafe fn eval_module_ctx( context: *mut JSContext, script: Script, ) -> Result<QuickJsValueAdapter, JsError>

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn report_ex(&self, err: &str) -> JSValue

throw an internal error to quickjs and create a new ex obj

pub unsafe fn report_ex_ctx(context: *mut JSContext, err: &str) -> JSValue

throw an Error in the runtime and init an Exception JSValue to return

Safety

when passing a context ptr please be sure that the corresponding QuickJsContext is still active

pub fn get_exception_ctx(&self) -> Option<JsError>

Get the last exception from the runtime, and if present, convert it to a JsError.

pub unsafe fn get_exception(context: *mut JSContext) -> Option<JsError>

Get the last exception from the runtime, and if present, convert it to a JsError.

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid

pub fn cache_object(&self, obj: QuickJsValueAdapter) -> i32

pub fn remove_cached_obj_if_present(&self, id: i32)

pub fn consume_cached_obj(&self, id: i32) -> QuickJsValueAdapter

pub fn with_cached_obj<C, R>(&self, id: i32, consumer: C) -> R
where C: FnOnce(QuickJsValueAdapter) -> R,

pub unsafe fn with_context<C, R>(context: *mut JSContext, consumer: C) -> R
where C: FnOnce(&QuickJsRealmAdapter) -> R,

Safety

When passing a context pointer please make sure the corresponding QuickJsContext is still valid

impl QuickJsRealmAdapter

pub fn get_realm_id(&self) -> &str

pub fn get_runtime_facade_inner(&self) -> Weak<QuickjsRuntimeFacadeInner>

pub fn get_script_or_module_name(&self) -> Result<String, JsError>

pub fn install_proxy( &self, proxy: JsProxy, add_global_var: bool, ) -> Result<QuickJsValueAdapter, JsError>

pub fn instantiate_proxy_with_id( &self, namespace: &[&str], class_name: &str, instance_id: usize, ) -> Result<QuickJsValueAdapter, JsError>

pub fn instantiate_proxy( &self, namespace: &[&str], class_name: &str, arguments: &[QuickJsValueAdapter], ) -> Result<(JsProxyInstanceId, QuickJsValueAdapter), JsError>

pub fn dispatch_proxy_event( &self, namespace: &[&str], class_name: &str, proxy_instance_id: &usize, event_id: &str, event_obj: &QuickJsValueAdapter, ) -> Result<bool, JsError>

pub fn dispatch_static_proxy_event( &self, namespace: &[&str], class_name: &str, event_id: &str, event_obj: &QuickJsValueAdapter, ) -> Result<bool, JsError>

pub fn install_function( &self, namespace: &[&str], name: &str, js_function: fn(&QuickJsRuntimeAdapter, &Self, &QuickJsValueAdapter, &[QuickJsValueAdapter]) -> Result<QuickJsValueAdapter, JsError>, arg_count: u32, ) -> Result<(), JsError>

pub fn install_closure<F: Fn(&QuickJsRuntimeAdapter, &Self, &QuickJsValueAdapter, &[QuickJsValueAdapter]) -> Result<QuickJsValueAdapter, JsError> + 'static>( &self, namespace: &[&str], name: &str, js_function: F, arg_count: u32, ) -> Result<(), JsError>

pub fn get_global(&self) -> Result<QuickJsValueAdapter, JsError>

pub fn get_namespace( &self, namespace: &[&str], ) -> Result<QuickJsValueAdapter, JsError>

pub fn invoke_function_on_object_by_name( &self, this_obj: &QuickJsValueAdapter, method_name: &str, args: &[QuickJsValueAdapter], ) -> Result<QuickJsValueAdapter, JsError>

pub fn invoke_function( &self, this_obj: Option<&QuickJsValueAdapter>, function_obj: &QuickJsValueAdapter, args: &[&QuickJsValueAdapter], ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_function<F: Fn(&Self, &QuickJsValueAdapter, &[QuickJsValueAdapter]) -> Result<QuickJsValueAdapter, JsError> + 'static>( &self, name: &str, js_function: F, arg_count: u32, ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_function_async<R, F>( &self, name: &str, js_function: F, arg_count: u32, ) -> Result<QuickJsValueAdapter, JsError>
where Self: Sized + 'static, R: Future<Output = Result<JsValueFacade, JsError>> + Send + 'static, F: Fn(JsValueFacade, Vec<JsValueFacade>) -> R + 'static,

pub fn create_error( &self, name: &str, message: &str, stack: &str, ) -> Result<QuickJsValueAdapter, JsError>

pub fn delete_object_property( &self, object: &QuickJsValueAdapter, property_name: &str, ) -> Result<(), JsError>

pub fn set_object_property( &self, object: &QuickJsValueAdapter, property_name: &str, property: &QuickJsValueAdapter, ) -> Result<(), JsError>

pub fn get_object_property( &self, object: &QuickJsValueAdapter, property_name: &str, ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_object(&self) -> Result<QuickJsValueAdapter, JsError>

pub fn construct_object( &self, constructor: &QuickJsValueAdapter, args: &[&QuickJsValueAdapter], ) -> Result<QuickJsValueAdapter, JsError>

pub fn get_object_properties( &self, object: &QuickJsValueAdapter, ) -> Result<Vec<String>, JsError>

pub fn traverse_object<F, R>( &self, object: &QuickJsValueAdapter, visitor: F, ) -> Result<Vec<R>, JsError>
where F: Fn(&str, &QuickJsValueAdapter) -> Result<R, JsError>,

pub fn traverse_object_mut<F>( &self, object: &QuickJsValueAdapter, visitor: F, ) -> Result<(), JsError>
where F: FnMut(&str, &QuickJsValueAdapter) -> Result<(), JsError>,

pub fn get_array_element( &self, array: &QuickJsValueAdapter, index: u32, ) -> Result<QuickJsValueAdapter, JsError>

pub fn push_array_element( &self, array: &QuickJsValueAdapter, element: &QuickJsValueAdapter, ) -> Result<u32, JsError>

push an element into an Array

pub fn set_array_element( &self, array: &QuickJsValueAdapter, index: u32, element: &QuickJsValueAdapter, ) -> Result<(), JsError>

pub fn get_array_length( &self, array: &QuickJsValueAdapter, ) -> Result<u32, JsError>

pub fn create_array(&self) -> Result<QuickJsValueAdapter, JsError>

pub fn traverse_array<F, R>( &self, array: &QuickJsValueAdapter, visitor: F, ) -> Result<Vec<R>, JsError>
where F: Fn(u32, &QuickJsValueAdapter) -> Result<R, JsError>,

pub fn traverse_array_mut<F>( &self, array: &QuickJsValueAdapter, visitor: F, ) -> Result<(), JsError>
where F: FnMut(u32, &QuickJsValueAdapter) -> Result<(), JsError>,

pub fn create_null(&self) -> Result<QuickJsValueAdapter, JsError>

pub fn create_undefined(&self) -> Result<QuickJsValueAdapter, JsError>

pub fn create_i32(&self, val: i32) -> Result<QuickJsValueAdapter, JsError>

pub fn create_string(&self, val: &str) -> Result<QuickJsValueAdapter, JsError>

pub fn create_boolean(&self, val: bool) -> Result<QuickJsValueAdapter, JsError>

pub fn create_f64(&self, val: f64) -> Result<QuickJsValueAdapter, JsError>

pub fn create_promise(&self) -> Result<QuickJsPromiseAdapter, JsError>

pub fn add_promise_reactions( &self, promise: &QuickJsValueAdapter, then: Option<QuickJsValueAdapter>, catch: Option<QuickJsValueAdapter>, finally: Option<QuickJsValueAdapter>, ) -> Result<(), JsError>

pub fn cache_promise(&self, promise_ref: QuickJsPromiseAdapter) -> usize

pub fn consume_cached_promise(&self, id: usize) -> Option<QuickJsPromiseAdapter>

pub fn dispose_cached_object(&self, id: i32)

pub fn with_cached_object<C, R>(&self, id: i32, consumer: C) -> R
where C: FnOnce(&QuickJsValueAdapter) -> R,

pub fn consume_cached_object(&self, id: i32) -> QuickJsValueAdapter

pub fn is_instance_of( &self, object: &QuickJsValueAdapter, constructor: &QuickJsValueAdapter, ) -> bool

pub fn json_stringify( &self, object: &QuickJsValueAdapter, opt_space: Option<&str>, ) -> Result<String, JsError>

pub fn json_parse( &self, json_string: &str, ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_typed_array_uint8( &self, buffer: Vec<u8>, ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_typed_array_uint8_copy( &self, buffer: &[u8], ) -> Result<QuickJsValueAdapter, JsError>

pub fn detach_typed_array_buffer( &self, array: &QuickJsValueAdapter, ) -> Result<Vec<u8>, JsError>

pub fn copy_typed_array_buffer( &self, array: &QuickJsValueAdapter, ) -> Result<Vec<u8>, JsError>

pub fn get_proxy_instance_info( &self, obj: &QuickJsValueAdapter, ) -> Result<(String, JsProxyInstanceId), JsError>
where Self: Sized,

pub fn to_js_value_facade( &self, js_value: &QuickJsValueAdapter, ) -> Result<JsValueFacade, JsError>
where Self: Sized + 'static,

pub fn from_js_value_facade( &self, value_facade: JsValueFacade, ) -> Result<QuickJsValueAdapter, JsError>
where Self: Sized + 'static,

convert a JSValueFacade into a JSValueAdapter you need this to move values into the worker thread from a different thread (JSValueAdapter cannot leave the worker thread)

pub fn value_adapter_to_serde_value( &self, value_adapter: &QuickJsValueAdapter, ) -> Result<Value, JsError>

pub fn serde_value_to_value_adapter( &self, value: Value, ) -> Result<QuickJsValueAdapter, JsError>

pub fn create_resolving_promise_async<P, R: Send + 'static, M>( &self, producer: P, mapper: M, ) -> Result<QuickJsValueAdapter, JsError>
where P: Future<Output = Result<R, JsError>> + Send + 'static, M: FnOnce(&QuickJsRealmAdapter, R) -> Result<QuickJsValueAdapter, JsError> + Send + 'static, Self: Sized + 'static,

create a new Promise with a Future which will run async and then resolve or reject the promise the mapper is used to convert the result of the future into a JSValueAdapter

pub fn create_resolving_promise<P, R: Send + 'static, M>( &self, producer: P, mapper: M, ) -> Result<QuickJsValueAdapter, JsError>
where P: FnOnce() -> Result<R, JsError> + Send + 'static, M: FnOnce(&QuickJsRealmAdapter, R) -> Result<QuickJsValueAdapter, JsError> + Send + 'static, Self: Sized + 'static,

create a new Promise with a FnOnce producer which will run async and then resolve or reject the promise the mapper is used to convert the result of the future into a JSValueAdapter