quickjs_runtime

Module reflection

utils for implementing proxy classes which can be used to use rust structs from JS (define method/getters/setters/etc)

Modules

eventtarget

EventTarget utils

Structs

Proxy

The Proxy struct can be used to create a class in JavaScript who’s methods can be implemented in rust

Constants

PROXY_INSTANCE_CLASS_ID

PROXY_STATIC_CLASS_ID

Functions

get_proxy

get a proxy by class_name (namespace.ClassName)

get_proxy_instance_id⚠

Get the instance id of a proxy instance

get_proxy_instance_id_q

get_proxy_instance_proxy_and_instance_id_q

is_proxy_instance⚠

check if an object is an instance of a Proxy class

is_proxy_instance_q

new_instance

new_instance2

Type Aliases

JsProxyInstanceId

ProxyCatchAllGetter

ProxyCatchAllSetter

ProxyConstructor

ProxyFinalizer

ProxyGetter

ProxyMethod

ProxyNativeMethod

ProxySetter

ProxyStaticCatchAllGetter

ProxyStaticCatchAllSetter

ProxyStaticGetter

ProxyStaticMethod

ProxyStaticNativeMethod

ProxyStaticSetter